requirements.txt
click==8.0.4
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
python-dotenv==0.19.2
Flask-WTF==1.0.0
Flask-sqlalchemy==2.5.1
Flask-migrate==3.1.0
Flask-login==0.5.0
email-validator==1.1.3
Flask-mail==0.9.1
pyjwt==2.3.0
Flask-bootstrap==3.3.7.1